+++ src/App.js
@@ -1,38 +1,27 @@
 import React from 'react'
-import {Route, Switch, withRouter, Link} from "react-router-dom"
-import {Trans, withNamespaces} from 'react-i18next'
-import LanguageChooser from './LanguageChooser'
-import TestStoreContext from './TestStoreContext'
-import NotFound from './NotFound'
+import {Route, Switch, withRouter} from "react-router-dom"
+import NotFound from './components/NotFound'
 import {routes} from './lib/router'
-import {Typography, Button} from '@material-ui/core'
-import Landing from './Landing'
+import {withTheme} from '@material-ui/core'
+import Landing from './components/Landing'
 
 import OfflineStatus from './components/OfflineStatus'
+import Register from './components/Register'
+import Profile from './components/Profile'
+import Dashboard from './components/Dashboard'
+import Settings from './components/Settings'
 
-const App = ({t}) =>
-  <div className="App">
+const App = ({theme}) =>
+  <div style={{backgroundColor: theme.palette.secondary.contrastText}}>
     <Switch>
       <Route component={Landing} exact path={routes.ROOT}/>
-      <Route component={() =>
-        <header className="App-header">
-          <Button color="primary" component={Link} to="404" variant="outlined">
-            {t("not-found.go-to")}
-          </Button>
-          <TestStoreContext/>
-          <Typography variant="h3">
-            {t('welcome.title', {framework: "react-i18next"})}
-          </Typography>
-          <Trans i18nKey="welcome.intro">
-            To get started, edit <code>src/App.js</code> and save to reload.
-          </Trans>
-          <LanguageChooser/>
-        </header>
-      } exact path={routes.TEST}
-      />
+      <Route component={Settings} exact path={routes.SETTINGS}/>
+      <Route component={Register} exact path={routes.REGISTER}/>
+      <Route component={Profile} exact path={routes.PROFILE}/>
+      <Route component={Dashboard} exact path={routes.DASHBOARD}/>
       <Route component={NotFound}/>
     </Switch>
     <OfflineStatus/>
   </div>
 
-export default withRouter(withNamespaces('common')(App))
+export default withRouter(withTheme()(App))