+++ src/db/__tests__/Store.test.js
@@ -1,0 +1,11 @@
+import {Database} from "../Store"
+
+describe("Store component", () => {
+
+  const wrapper = shallow(<Database/>)
+
+  it("renders correctly", () => {
+    expect(wrapper).toHaveLength(1)
+  })
+
+})