+++ src/index.js
@@ -4,9 +4,9 @@ import {BrowserRouter as Router} from "react-router-dom"
 import {MuiThemeProvider} from '@material-ui/core/styles'
 import App from './App'
 import "./lib/firebase"
-import theme from "./lib/material-ui"
+import {light, dark} from "./lib/material-ui"
 import "./main.sass"
-import {Database} from './db'
+import Store, {Database} from './db'
 
 import i18next from "./lib/i18next"
 import {I18nextProvider} from 'react-i18next'
@@ -17,9 +17,13 @@ ReactDOM.render(
   <I18nextProvider i18n={i18next}>
     <Database>
       <Router>
-        <MuiThemeProvider theme={theme}>
-          <App/>
-        </MuiThemeProvider>
+        <Store.Consumer>
+          {({isDarkMode}) =>
+            <MuiThemeProvider theme={isDarkMode ? dark : light}>
+              <App/>
+            </MuiThemeProvider>
+          }
+        </Store.Consumer>
       </Router>
     </Database>
   </I18nextProvider>,