+++ src/lib/i18next.js
@@ -1,21 +1,41 @@
 import i18next from 'i18next'
 import common_en from "../locales/en/common.json"
 import common_no from "../locales/no/common.json"
+import profile_en from "../locales/en/profile.json"
+import profile_no from "../locales/no/profile.json"
 import pages_en from "../locales/en/pages.json"
 import pages_no from "../locales/no/pages.json"
+import settings_en from "../locales/en/settings.json"
+import settings_no from "../locales/no/settings.json"
 import languages from "../locales/locales.json"
 
 
 i18next.init({
-  lng: navigator.language || languages.default,
+  lng: localStorage.getItem("preferredLanguage") || navigator.language || languages.default,
   resources: {
     en: {
       common: common_en,
-      pages: pages_en
+      profile: profile_en,
+      pages: pages_en,
+      settings: settings_en
     },
     no: {
       common: common_no,
-      pages: pages_no
+      profile: profile_no,
+      pages: pages_no,
+      settings: settings_no
+    },
+    "ny-NO":{
+      common: common_no,
+      pages: pages_no,
+      profile: profile_no,
+      settings: settings_no
+    },
+    "nb-NO": {
+      common: common_no,
+      pages: pages_no,
+      profile: profile_no,
+      settings: settings_no
     }
   }
 })