+++ CHANGELOG.md
@@ -1,8 +1,24 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [Unreleased]
- - User now gets a notification that the PWA can be added to homescreen
+## [Unreleased] 
+### Added
+- User now gets a notification that the PWA can be added to homescreen
+- use `localStorage` to save preferred language over sessions
+- settings page
+- dark mode
+- authentication:
+    - registration
+    - login
+    - logout
+    - user deletion
+    - redirects:
+        - logged out
+            - after registering a user -> dashboard
+            - dashboard -> root
+            - 404 -> root
+### Removed
+- example codes
 
 ## [0.3.0] - 2019-01-30
 ### Added